# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smiley API
  version: 0.1.0
servers:
  - url: http://localhost:3000/api/v1
paths:
  /auth/register:
    post:
      summary: Register a user
      responses:
        "201":
          description: Created
  /auth/login:
    post:
      summary: Login
      responses:
        "200":
          description: Auth session
  /auth/password-reset/request:
    post:
      summary: Request password reset
      responses:
        "202":
          description: Accepted
  /users/search:
    get:
      summary: Search users by username
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Search results
  /partnership-requests:
    post:
      summary: Create a partnership request
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
    get:
      summary: List pending partnership requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Requests
  /partnership-requests/{id}/accept:
    post:
      summary: Accept a partnership request
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Partnership created
  /partnership-requests/{id}/reject:
    post:
      summary: Reject a partnership request
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Rejected
  /partnership-requests/{id}/cancel:
    post:
      summary: Cancel an outgoing partnership request
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Cancelled
  /partnerships/current:
    get:
      summary: Get current partnership
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current partnership or null
  /partnerships/{id}/onboarding:
    post:
      summary: Complete relationship onboarding
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Activated partnership
  /space:
    get:
      summary: Get active world summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: World summary
  /posts:
    get:
      summary: List shared posts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Posts
    post:
      summary: Create a shared post
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /moods:
    get:
      summary: List moods
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Moods
    post:
      summary: Create mood update
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /messages:
    get:
      summary: List conversation messages
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Messages
    post:
      summary: Send idempotent text message
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /calendar-events:
    get:
      summary: List calendar events
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Events
    post:
      summary: Create calendar event
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /notifications:
    get:
      summary: List notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
  /notifications/read-all:
    post:
      summary: Mark notifications as read
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Updated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT